Added static methods to EpsilonComparer. These should be used in the main codebase where comparisons are needed. This removes the need to do something like EpsilonComparer.DoubleComparer.Compare(x, y). Instead you can just do EpsilonComparer.Compare(x, y). Right now there's only a static Compare method for doubles, but we can add more overloads if/when they are needed.
EpsilonComparer.DoubleComparer and EpsilonComparer.VectorComparer are still there for cases when you need an actual instance of IComparer<T> (very useful for tests).

In general, lets try to keep comparisons in the tests and nowhere else. Vector.IsUnit is the only method in the main codebase that uses EpsilonComparer, and that property is only used in tests. We should probably move that method out to the test assembly...