TONS of perf optimizations.
We are now nearly 8x as fast on the simple visualizations I've been running (that's right, 800%).

BitmapRenderTarget has been deprecated because it's too slow. Use MemoryRenderTarget instead and use ToBitmap to convert it to a System.Drawing.Bitmap instance. THIS IS MUCH FASTER. Setting pixels doesn't even show up in the profiler anymore.

Optimized CrossProduct. It was multiplying with the unit vectors to obtain its result. While that's correct from a mathematical perspective, it's also unnecessary and redundant.